修复 Linux 下头文件大小写敏感，以及几处编译警告。

+++ fibjs/src/BufferedStream.cpp
@@ -285,8 +285,8 @@ result_t BufferedStream::readUntil(const char* mk, int32_t maxlen,
 			}
 
 			if (maxlen > 0
-					&& pThis->m_strbuf.size() + (pos - pThis->m_pos)
-							> maxlen + mklen)
+					&& ((int)pThis->m_strbuf.size() + (pos - pThis->m_pos)
+							> maxlen + mklen))
 				return CALL_E_INVALID_DATA;
 
 			pThis->append(pos - pThis->m_pos);
